% Auto-generated: do not edit by hand
\name{chart}

\alias{chart}

\title{Chart component}

\description{

}

\usage{
chart(id=NULL, option=NULL, config=NULL, style=NULL,
resizeObserver=NULL)
}

\arguments{
\item{id}{Logical | numeric | character | named list | unnamed list. }

\item{option}{Logical | numeric | character | named list | unnamed list. }

\item{config}{Logical | numeric | character | named list | unnamed list. }

\item{style}{Logical | numeric | character | named list | unnamed list. }

\item{resizeObserver}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

