% Auto-generated: do not edit by hand
\name{dashECharts}

\alias{dashECharts}

\title{DashECharts component}

\description{

}

\usage{
dashECharts(id=NULL, brush_data=NULL, events=NULL, fun_effects=NULL,
fun_keys=NULL, fun_paths=NULL, fun_values=NULL, funs=NULL,
lazyUpdate=NULL, maps=NULL, n_clicks=NULL,
n_clicks_data=NULL, n_clicks_timestamp=NULL, notMerge=NULL,
option=NULL, opts=NULL, selected_data=NULL, style=NULL,
theme=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{brush_data}{Named list. }

\item{events}{Unnamed list. }

\item{fun_effects}{Unnamed list. }

\item{fun_keys}{Unnamed list. }

\item{fun_paths}{Named list. }

\item{fun_values}{Unnamed list. }

\item{funs}{Named list. }

\item{lazyUpdate}{Logical. }

\item{maps}{Named list. }

\item{n_clicks}{Numeric. }

\item{n_clicks_data}{Named list. }

\item{n_clicks_timestamp}{Numeric. }

\item{notMerge}{Logical. }

\item{option}{Named list. }

\item{opts}{Named list. }

\item{selected_data}{Named list. }

\item{style}{Named list. }

\item{theme}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

