% Auto-generated: do not edit by hand
\name{dashECharts}

\alias{dashECharts}

\title{DashECharts component}

\description{

}

\usage{
dashECharts(n_clicks=NULL, n_clicks_timestamp=NULL, n_clicks_data=NULL,
selected_data=NULL, brush_data=NULL, option=NULL,
notMerge=NULL, lazyUpdate=NULL, theme=NULL, events=NULL,
style=NULL, opts=NULL, maps=NULL, fun_formatter=NULL,
id=NULL)
}

\arguments{
\item{n_clicks}{Numeric. }

\item{n_clicks_timestamp}{Numeric. }

\item{n_clicks_data}{Named list. }

\item{selected_data}{Named list. }

\item{brush_data}{Named list. }

\item{option}{Named list. }

\item{notMerge}{Logical. }

\item{lazyUpdate}{Logical. }

\item{theme}{Character. }

\item{events}{Unnamed list. }

\item{style}{Named list. }

\item{opts}{Named list. }

\item{maps}{Named list. }

\item{fun_formatter}{Logical. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

