% Auto-generated: do not edit by hand
\name{dashECharts}

\alias{dashECharts}

\title{DashECharts component}

\description{

}

\usage{
dashECharts(id=NULL, bmap_token=NULL, brush_data=NULL, click_data=NULL,
event=NULL, fun_effects=NULL, fun_keys=NULL, fun_paths=NULL,
fun_prepares=NULL, fun_values=NULL, funs=NULL,
mapbox_token=NULL, maps=NULL, n_clicks=NULL,
n_clicks_timestamp=NULL, option=NULL, reset_id=NULL,
resize_id=NULL, selected_data=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{bmap_token}{Character. }

\item{brush_data}{Named list. }

\item{click_data}{Named list. }

\item{event}{Named list. }

\item{fun_effects}{Unnamed list. }

\item{fun_keys}{Unnamed list. }

\item{fun_paths}{Named list. }

\item{fun_prepares}{Unnamed list. }

\item{fun_values}{Unnamed list. }

\item{funs}{Named list. }

\item{mapbox_token}{Character. }

\item{maps}{Named list. }

\item{n_clicks}{Numeric. }

\item{n_clicks_timestamp}{Numeric. }

\item{option}{Named list. }

\item{reset_id}{Numeric. }

\item{resize_id}{Numeric. }

\item{selected_data}{Named list. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

